refactor code to improve organization, fixes
