+++ jaims/__init__.py
@@ -1,13 +1,9 @@
 from .agent import JAImsAgent
 from .openai_wrappers import (
-    JAImsGPTModel,
-    JAImsTokensExpense,
     estimate_token_count,
-    JAImsOpenaiKWArgs,
-    JAImsOptions,
 )
 
-# TODO: remove stars and export only what's necessary
-from .exceptions import *
-from .function_handler import *
+
+from .function_handler import JAImsToolHandler
 from .transaction_storage import JAImsTransactionStorageInterface
+from .entities import *
