+++ jaims/agent.py
@@ -1,7 +1,7 @@
 import json
 import logging
 import os
-from typing import Any, Generator, List, Optional, Union
+from typing import Generator, List, Optional, Union
 
 from jaims.transaction_storage import (
     JAImsTransactionStorageInterface,
@@ -11,60 +11,23 @@ import openai
 from openai.types.chat import ChatCompletionChunk, ChatCompletion
 
 from jaims.openai_wrappers import (
-    JAImsGPTModel,
-    JAImsTokensExpense,
-    JAImsOpenaiKWArgs,
-    JAImsOptions,
     estimate_token_count,
     get_openai_response,
 )
-from jaims.exceptions import (
-    JAImsMissingOpenaiAPIKeyException,
-    JAImsMaxConsecutiveFunctionCallsExceeded,
-)
+
 from jaims.function_handler import (
-    JAImsFunctionHandler,
+    JAImsToolHandler,
 )
 from jaims.history_manager import HistoryManager
 
-
-class JAImsCallContext:
-    """
-    Represents the context for a JAIms run.
-    It is used when in case of function calling, the run restarts recursively.
-
-    Args:
-        openai_kwargs (JAImsOpenaiKWArgs): The OpenAI keyword arguments.
-        options (JAImsOptions): The JAIms options.
-        iterations (int, optional): The number of iterations. Defaults to 0.
-    """
-
-    def __init__(
-        self,
-        openai_kwargs: JAImsOpenaiKWArgs,
-        options: JAImsOptions,
-        iterations: int = 0,
-    ):
-        self.openai_kwargs = openai_kwargs
-        self.options = options
-        self.iterations = iterations
-
-    def add_iteration(self):
-        self.iterations += 1
-        if self.iterations > self.options.max_consecutive_function_calls:
-            raise JAImsMaxConsecutiveFunctionCallsExceeded(
-                f"Max consecutive function calls exceeded ({self.options.max_consecutive_function_calls})"
-            )
-
-    def update(
-        self,
-        openai_kwargs: Optional[JAImsOpenaiKWArgs] = None,
-        options: Optional[JAImsOptions] = None,
-    ):
-        if openai_kwargs:
-            self.openai_kwargs = openai_kwargs
-        if options:
-            self.options = options
+from jaims.entities import (
+    JAImsGPTModel,
+    JAImsOpenaiKWArgs,
+    JAImsOptions,
+    JAImsTokensExpense,
+    JAImsMissingOpenaiAPIKeyException,
+    JAImsMaxConsecutiveFunctionCallsExceeded,
+)
 
 
 class JAImsAgent:
@@ -103,19 +66,50 @@ class JAImsAgent:
         self.__options = options
         self.__expense = JAImsAgent.__init_expense_dictionary()
         self.__last_run_expense = JAImsAgent.__init_expense_dictionary()
-        self.__function_handler = JAImsFunctionHandler()
+        self.__function_handler = JAImsToolHandler()
         self.__history_manager = HistoryManager()
+
+        if self.__openai_kwargs.messages:
+            self.__history_manager.add_messages(self.__openai_kwargs.messages)
+
+        self.__iteration_count = 0
         if transaction_storage is None:
             self.__transaction_storage = JAImsTransactionStorageInterface()
         else:
             self.__transaction_storage = transaction_storage
 
+    def __increment_iteration_count(self):
+        self.__iteration_count += 1
+        if self.__iteration_count > self.__options.max_consecutive_function_calls:
+            raise JAImsMaxConsecutiveFunctionCallsExceeded(
+                f"Max consecutive function calls exceeded ({self.__options.max_consecutive_function_calls})"
+            )
+
+    def __reset_iteration_count(self):
+        self.__iteration_count = -1
+
+    def set_openai_kwargs(self, openai_kwargs: JAImsOpenaiKWArgs):
+        """
+        Updates the openai kwargs of the agent.
+
+        Args:
+            openai_kwargs (JAImsOpenaiKWArgs): The new openai kwargs.
+        """
+        self.__openai_kwargs = openai_kwargs
+
+    def set_options(self, options: JAImsOptions):
+        """
+        Updates the options of the agent.
+
+        Args:
+            options (JAImsOptions): The new options.
+        """
+        self.__options = options
+
     # TODO: add support for multiple responses when n>1
     def run(
         self,
         messages: Optional[List[dict]] = None,
-        override_options: Optional[JAImsOptions] = None,
-        override_openai_kwargs: Optional[JAImsOpenaiKWArgs] = None,
     ) -> Union[str, Generator[str, None, None]]:
         """
         Starts a run calling openai with the passed messages.
@@ -123,47 +117,34 @@ class JAImsAgent:
 
         Args:
             messages (Optional[List[dict]], optional): The messages to be sent to OpenAI.
-            override_options (Optional[JAImsOptions], optional): The options to be used for this run (entirely overriding those passed in constructor, only for this run).
-            override_openai_kwargs (Optional[JAImsOpenaiKWArgs], optional): The OpenAI keyword arguments to be used for this run (entirely overriding those passed in constructor, only for this run).
         """
+        self.__reset_iteration_count()
         messages = messages or []
-        if override_openai_kwargs:
-            messages.extend(override_openai_kwargs.messages or [])
-
         self.__history_manager.add_messages(messages)
-        options = override_options or self.__options
-        openai_kwargs = override_openai_kwargs or self.__openai_kwargs
-        call_context = JAImsCallContext(openai_kwargs, options)
 
         self.__last_run_expense = JAImsAgent.__init_expense_dictionary()
-        return self.__call_openai(call_context)
+        return self.__call_openai()
 
-    def __call_openai(
-        self, call_context: JAImsCallContext
-    ) -> Union[str, Generator[str, None, None]]:
+    def __call_openai(self) -> Union[str, Generator[str, None, None]]:
         # throws if max consecutive calls is exceeded
-        call_context.add_iteration()
+        self.__increment_iteration_count()
         messages = self.__history_manager.get_messages_for_current_run(
-            options=call_context.options,
-            openai_kwargs=call_context.openai_kwargs,
+            options=self.__options,
+            openai_kwargs=self.__openai_kwargs,
         )
 
-        call_context.openai_kwargs.messages = messages
+        self.__openai_kwargs.messages = messages
 
-        response = get_openai_response(
-            call_context.openai_kwargs,
-            call_context.options,
-        )
+        response = get_openai_response(self.__openai_kwargs, self.__options)
 
         if isinstance(response, openai.Stream):
-            return self.__handle_streaming_response(response, call_context)
+            return self.__handle_streaming_response(response)
         else:
-            return self.__handle_response(response, call_context)
+            return self.__handle_response(response)
 
     def __handle_streaming_response(
         self,
         response: openai.Stream[ChatCompletionChunk],
-        call_context: JAImsCallContext,
     ) -> Generator[str, None, None]:
         accumulated_chunks = None
         for completion_chunk in response:
@@ -175,7 +156,7 @@ class JAImsAgent:
                 )
 
                 if (
-                    call_context.options.debug_stream_function_call
+                    self.__options.debug_stream_function_call
                     and message_delta.tool_calls
                 ):
                     print(
@@ -186,43 +167,37 @@ class JAImsAgent:
 
                 if completion_chunk.choices[0].finish_reason is not None:
                     if (
-                        call_context.options.debug_stream_function_call
+                        self.__options.debug_stream_function_call
                         and accumulated_chunks.tool_calls
                     ):
                         print("\n")
 
                     # rebuilding entire completion chunk with accumulated delta
                     completion_chunk.choices[0].delta = accumulated_chunks
-                    self.__store_transaction(call_context, completion_chunk)
+                    self.__store_transaction(completion_chunk)
 
                     self.__handle_token_expense_from_openai_response(
-                        accumulated_chunks.model_dump(), call_context
+                        accumulated_chunks.model_dump()
                     )
 
-                    yield from self.__handle_response_message(
-                        accumulated_chunks, call_context
-                    )
+                    yield from self.__evaluate_next_run_from_message(accumulated_chunks)
 
                 if message_delta.content:
                     yield message_delta.content
 
-    def __handle_response(
-        self, response: ChatCompletion, call_context: JAImsCallContext
-    ):
+    def __handle_response(self, response: ChatCompletion):
         if len(response.choices) == 0:
             return ""
 
         # log token expense
-        self.__handle_token_expense_from_openai_response(
-            response.model_dump(), call_context
-        )
+        self.__handle_token_expense_from_openai_response(response.model_dump())
 
-        self.__store_transaction(call_context, response)
+        self.__store_transaction(response)
 
         message = response.choices[0].message
-        return self.__handle_response_message(message, call_context)
+        return self.__evaluate_next_run_from_message(message)
 
-    def __handle_response_message(self, message, call_context: JAImsCallContext):
+    def __evaluate_next_run_from_message(self, message):
         logger = logging.getLogger(__name__)
         logger.debug(f"OpenAI response message:\n{message}")
 
@@ -230,14 +205,20 @@ class JAImsAgent:
         self.__history_manager.add_messages([message_dict])
 
         if message.tool_calls:
-            tools_results = self.__function_handler.handle_from_message(
-                message=message_dict,
-                function_wrappers=call_context.openai_kwargs.tools or [],
+            tools_results = self.__function_handler.handle_tool_calls(
+                tool_calls=message_dict.get("tool_calls", []),
+                function_wrappers=self.__openai_kwargs.tools or [],
             )
+
+            if tools_results.override_kwargs:
+                self.set_openai_kwargs(tools_results.override_kwargs)
+
+            if tools_results.override_options:
+                self.set_options(tools_results.override_options)
 
             if (
                 tools_results.stop == False
-                and call_context.openai_kwargs.tool_choice == "auto"
+                and self.__openai_kwargs.tool_choice == "auto"
             ):
                 messages = tools_results.function_result_messages
                 messages.extend(
@@ -245,52 +226,45 @@ class JAImsAgent:
                     if tools_results.override_kwargs is not None
                     else []
                 )
-                self.__history_manager.add_messages(message)
-                call_context.update(
-                    openai_kwargs=call_context.openai_kwargs,
-                    options=call_context.options,
-                )
-                return self.__call_openai(call_context)
+                self.__history_manager.add_messages(messages)
+                return self.__call_openai()
 
         # if the response was streaming only an empty string must be returned
         # because the streaming generator already yielded the response, otherwise
         # the content is returned.
-        if call_context.openai_kwargs.stream:
+        if self.__openai_kwargs.stream:
             return ""
         else:
             return message.content or ""
 
     def __store_transaction(
         self,
-        call_context: JAImsCallContext,
         response: Union[ChatCompletion, ChatCompletionChunk],
     ):
         self.__transaction_storage.store_transaction(
-            request=call_context.openai_kwargs.to_dict(),
+            request=self.__openai_kwargs.to_dict(),
             response=response.model_dump(exclude_none=True),
         )
 
-    def __handle_token_expense_from_openai_response(
-        self, response, call_context: JAImsCallContext
-    ):
-        if not call_context.openai_kwargs.stream:
+    def __handle_token_expense_from_openai_response(self, response):
+        if not self.__openai_kwargs.stream:
             expense = JAImsTokensExpense.from_openai_usage_dictionary(
-                call_context.openai_kwargs.model, response["usage"]
+                self.__openai_kwargs.model, response["usage"]
             )
         else:
             sent_messages = self.__history_manager.get_messages_for_current_run(
-                options=call_context.options,
-                openai_kwargs=call_context.openai_kwargs,
+                options=self.__options,
+                openai_kwargs=self.__openai_kwargs,
             )
             prompt_tokens = estimate_token_count(
-                json.dumps(sent_messages), model=call_context.openai_kwargs.model
+                json.dumps(sent_messages), model=self.__openai_kwargs.model
             )
             completion_tokens = estimate_token_count(
-                json.dumps(response), model=call_context.openai_kwargs.model
+                json.dumps(response), model=self.__openai_kwargs.model
             )
             total_tokens = prompt_tokens + completion_tokens
             expense = JAImsTokensExpense(
-                gpt_model=call_context.openai_kwargs.model,
+                gpt_model=self.__openai_kwargs.model,
                 prompt_tokens=prompt_tokens,
                 completion_tokens=completion_tokens,
                 total_tokens=total_tokens,
